# Conflicts:
#	pom.xml
#	src/Components/Splash/SplashController.java
#	src/Reports/ProjectsReport.java
#	src/Reports/TasksReport.java
#	src/Reports/UsersReport.java
#	src/Service/Database.java

+++ src/model/UsersDAO.java
@@ -1,5 +1,6 @@
 package model;
 
+import Service.Session;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import Utilize.DatabaseHandler;
@@ -87,7 +88,28 @@ public class UsersDAO {
         }
     }
 
-    public static ObservableList<Users> searchUsers (String data) throws SQLException, ClassNotFoundException {
+    public static Users searchUsers (String data) throws SQLException, ClassNotFoundException {
+
+        String selectStmt = "SELECT * " +
+                "FROM users " +
+                "WHERE display_name = '"+ data +"';";
+
+        try {
+            ResultSet rsUsers = DatabaseHandler.databaseExecuteQuery(selectStmt);
+
+            Users user = getUserFromResultSet(rsUsers);
+
+            return user;
+
+        } catch (Exception e) {
+            System.out.println("Wystąpił błąd podczas wyszukiwania użytkowników " + e);
+            e.printStackTrace();
+
+            return null;
+        }
+    }
+
+    /*public static ObservableList<Users> searchUsers (String data) throws SQLException, ClassNotFoundException {
 
         String selectStmt = "SELECT * " +
                 "FROM users " +
@@ -106,6 +128,18 @@ public class UsersDAO {
             e.printStackTrace();
 
             return null;
+        }
+    }*/
+
+    public static void updateUser (String email, String password, String displayName) throws  SQLException, ClassNotFoundException {
+
+        String updateStm =
+                    "UPDATE Users SET email= '"+email+"', password='"+password+"', display_name='"+displayName+"' WHERE user_id="+Session.getUserId();
+        try {
+            DatabaseHandler.databaseExecuteUpdate(updateStm);
+        }catch (Exception e){
+            System.out.print("Nie udało się edytować użytkownika, nie patrz na stack trace bo dostaniesz raka " + e);
+            e.printStackTrace();
         }
     }
 
@@ -128,8 +162,7 @@ public class UsersDAO {
     public static void deleteUser (int id) throws SQLException, ClassNotFoundException {
 
         String updateStmt =
-                "DELETE FROM users" +
-                        "WHERE user_id =" + id;
+                "DELETE FROM users WHERE user_id =" + id;
         try {
             DatabaseHandler.databaseExecuteUpdate(updateStmt);
         } catch (SQLException e) {
